+++ projects/ERM19264_UC1609C/Core/Inc/custom_graphics.h
@@ -1,101 +1,114 @@
 /*
-* Project Name: ERM19264_UC1609
-* File: custom_graphics.h
-* Description: ERM19264 LCD driven by UC1609C controller header file for the custom graphics functions based on Adafruit graphics library
-* Author: Gavin Lyons.
-*/
+ * Project Name: ERM19264_UC1609
+ * File: custom_graphics.h
+ * Description: ERM19264 LCD driven by UC1609C controller header file for the custom graphics functions
+ */
 
 #ifndef _CUSTOM_GRAPHICS_H
 #define _CUSTOM_GRAPHICS_H
 
 #include "Print.h"
 
-
 #define swap(a, b) { int16_t t = a; a = b; b = t; }
 
-#define ERM19264_ASCII_OFFSET 0x00
-#define ERM19264_ASCII_OFFSET_SP 0x20 // Starts at Space
-#define ERM19264_ASCII_OFFSET_NUM 0x30 // Starts at number 0
+typedef enum {
+	UC1609Font_Default = 1,
+	UC1609Font_Thick = 2,
+	UC1609Font_Seven_Seg = 3,
+	UC1609Font_Wide = 4,
+	UC1609Font_Bignum = 5,
+	UC1609Font_Mednum = 6,
+} LCD_FONT_TYPE_e;
 
-class custom_graphics : public Print {
+typedef enum {
+	FONT_W_5 = 5, FONT_W_7 = 7, FONT_W_4 = 4, FONT_W_8 = 8, FONT_W_16 = 16
+} LCD_Font_width_e; // width of the font in bytes cols.
 
- public:
+typedef enum {
+	FONT_O_EXTEND = 0x00, //   extends ASCII
+	FONT_O_SP = 0x20,  // Starts at Space
+	FONT_O_NUM = 0x30,  // Starts at number '0'
+} LCD_Font_offset_e; // font offset in the ASCII table
 
-  custom_graphics(int16_t w, int16_t h); // Constructor
+typedef enum {
+	FONT_H_8 = 8, FONT_H_16 = 16, FONT_H_32 = 32
+} LCD_Font_height_e; // width of the font in bits
 
-  // This MUST be defined by the subclass:
-  virtual void drawPixel(int16_t x, int16_t y, uint8_t color) = 0;
+class custom_graphics: public Print {
 
-  // These MAY be overridden by the subclass to provide device-specific
-  // optimized code.  Otherwise 'generic' versions are used.
-  virtual void
-    drawLine(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint8_t color),
-    drawFastVLine(int16_t x, int16_t y, int16_t h, uint8_t color),
-    drawFastHLine(int16_t x, int16_t y, int16_t w, uint8_t color),
-    drawRect(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t color),
-    fillRect(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t color),
-    fillScreen(uint8_t color),
-    invertDisplay(bool i);
+public:
 
-  // no subclass overrides
-  void
-    drawCircle(int16_t x0, int16_t y0, int16_t r, uint8_t color),
-    drawCircleHelper(int16_t x0, int16_t y0, int16_t r, uint8_t cornername,
-      uint8_t color),
-    fillCircle(int16_t x0, int16_t y0, int16_t r, uint8_t color),
-    fillCircleHelper(int16_t x0, int16_t y0, int16_t r, uint8_t cornername,
-      int16_t delta, uint8_t color),
-    drawTriangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1,
-      int16_t x2, int16_t y2, uint8_t color),
-    fillTriangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1,
-      int16_t x2, int16_t y2, uint8_t color),
-    drawRoundRect(int16_t x0, int16_t y0, int16_t w, int16_t h,
-      int16_t radius, uint8_t color),
-    fillRoundRect(int16_t x0, int16_t y0, int16_t w, int16_t h,
-      int16_t radius, uint8_t color),
-    drawBitmap(int16_t x, int16_t y, const uint8_t *bitmap,
-      int16_t w, int16_t h, uint8_t color),
-    drawBitmap(int16_t x, int16_t y, const uint8_t *bitmap,
-      int16_t w, int16_t h, uint8_t color, uint8_t bg),
-    drawXBitmap(int16_t x, int16_t y, const uint8_t *bitmap, 
-      int16_t w, int16_t h, uint8_t color),
-    drawChar(int16_t x, int16_t y, unsigned char c, uint8_t color,
-      uint8_t bg, uint8_t size),
-    setCursor(int16_t x, int16_t y),
-    setTextColor(uint8_t c),
-    setTextColor(uint8_t c, uint8_t bg),
-    setTextSize(uint8_t s),
-    setTextWrap(bool w),
-    setRotation(uint8_t r);
+	custom_graphics(int16_t w, int16_t h); // Constructor
 
-  virtual size_t write(uint8_t);
+	// This MUST be defined by the subclass:
+	virtual void drawPixel(int16_t x, int16_t y, uint8_t color) = 0;
 
-    void setFontNum(uint8_t FontNumber);
-    void drawCharNumFont(uint8_t x, uint8_t y, uint8_t c, uint8_t color , uint8_t bg);
-    void drawTextNumFont(uint8_t x, uint8_t y, char *pText, uint8_t color, uint8_t bg);
-  
-  int16_t height(void) const;
-  int16_t width(void) const;
+	void  drawLine(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint8_t color);
+	void drawFastVLine(int16_t x, int16_t y, int16_t h, uint8_t color);
+	void drawFastHLine(int16_t x, int16_t y, int16_t w, uint8_t color);
+	void drawRect(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t color);
+	void fillRect(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t color);
+	void fillScreen(uint8_t color);
 
-  uint8_t getRotation(void) const;
+	void drawCircle(int16_t x0, int16_t y0, int16_t r, uint8_t color);
+	void drawCircleHelper(int16_t x0, int16_t y0, int16_t r, uint8_t cornername,
+			uint8_t color);
+	void fillCircle(int16_t x0, int16_t y0, int16_t r, uint8_t color);
+	void fillCircleHelper(int16_t x0, int16_t y0, int16_t r, uint8_t cornername,
+			int16_t delta, uint8_t color);
+	void drawTriangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1,
+			int16_t x2, int16_t y2, uint8_t color);
+			void fillTriangle(int16_t x0, int16_t y0, int16_t x1, int16_t y1,
+			int16_t x2, int16_t y2, uint8_t color);
+	void drawRoundRect(int16_t x0, int16_t y0, int16_t w, int16_t h,
+			int16_t radius, uint8_t color);
+	void fillRoundRect(int16_t x0, int16_t y0, int16_t w, int16_t h,
+			int16_t radius, uint8_t color);
+	void drawBitmap(int16_t x, int16_t y, const uint8_t *bitmap,
+			int16_t w, int16_t h, uint8_t color, uint8_t bg);
+	void drawChar(int16_t x, int16_t y, unsigned char c, uint8_t color,
+			uint8_t bg, uint8_t size);
+	void setCursor(int16_t x, int16_t y);
+	void setTextColor(uint8_t c);
+	void setTextColor(uint8_t c, uint8_t bg);
+	void setTextSize(uint8_t s);
+	void setTextWrap(bool w);
+	void setRotation(uint8_t r);
 
- protected:
-  const int16_t
-    WIDTH, HEIGHT;   // This is the 'raw' display w/h - never changes
-  int16_t
-    _width, _height, // Display w/h as modified by current rotation
-    cursor_x, cursor_y;
-  uint8_t
-    textcolor, textbgcolor;
-  uint8_t
-    textsize,
-    rotation;
-  bool
-    wrap; // If set, 'wrap' text at right edge of display
-    uint8_t _FontNumber = 1;
-    uint8_t _CurrentFontWidth = 5;
-    uint8_t _CurrentFontoffset = 0;
-    uint8_t _CurrentFontheight = 8;
+	virtual size_t write(uint8_t);
+
+	void setFontNum(LCD_FONT_TYPE_e FontNumber);
+	void drawCharNumFont(uint8_t x, uint8_t y, uint8_t c, uint8_t color,
+			uint8_t bg);
+	void drawTextNumFont(uint8_t x, uint8_t y, char *pText, uint8_t color,
+			uint8_t bg);
+
+	void  setDrawBitmapAddr(bool mode);
+
+	int16_t height(void) const;
+	int16_t width(void) const;
+
+	uint8_t getRotation(void) const;
+	LCD_FONT_TYPE_e FontNum;
+
+protected:
+	const int16_t WIDTH, HEIGHT; // This is the 'raw' display w/h - never changes
+	int16_t _width;
+	int16_t  _height; // Display w/h as modified by current rotation
+	int16_t cursor_x;
+	int16_t cursor_y;
+	uint8_t textcolor;
+	uint8_t textbgcolor;
+
+	uint8_t textsize;
+	uint8_t rotation;
+	bool wrap; // If set, 'wrap' text at right edge of display
+	bool drawBitmapAddr; // True vertical , false = horizontal
+
+	uint8_t _FontNumber = UC1609Font_Default;
+	uint8_t _CurrentFontWidth = FONT_W_5;
+	uint8_t _CurrentFontoffset = FONT_O_EXTEND;
+	uint8_t _CurrentFontheight = FONT_H_8;
 };
 
 #endif