+++ projects/ERM19264_UC1609C/examples/HWSPI_NOBUFFER_main.cpp
@@ -7,6 +7,7 @@
 // *****************************
 // NOTES :
 // (1) This is for hardware SPI. 
+// (2) no buffer mode must be selected in llibrary header file
 // ******************************
  */